oss-arch-gym/sims/DRAM/requirement.txt
absl-py==1.4.0
ale-py==0.8.1
astunparse==1.6.3
atari-py==0.2.9
attrs==23.2.0
bsuite==0.3.5
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.6
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
distrax==0.1.3
dm-acme @ file:///home/user/Desktop/oss-arch-gym/copy/acme
dm-control==0.0.364896371
dm-env==1.6
dm-haiku==0.0.10
dm-launchpad==0.5.2
dm-reverb==0.7.2
dm-sonnet==2.0.2
dm-tree==0.1.8
envlogger==1.2
exceptiongroup==1.2.0
execnet==2.0.2
flatbuffers==24.3.25
fonttools==4.50.0
frozendict==2.4.1
future==1.0.0
gast==0.5.4
glfw==2.7.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlab==0.8.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
jax==0.4.3
jaxlib==0.4.3
jmp==0.0.4
joblib==1.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
labmaze==1.0.6
lazy_loader==0.3
libclang==18.1.1
libcst==1.2.0
lxml==5.2.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
mizani==0.10.0
mock==5.1.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.7
packaging==24.0
pandas==2.2.1
patsy==0.5.6
pillow==10.3.0
plotnine==0.10.1
pluggy==1.4.0
portpicker==1.6.0
promise==2.3
protobuf==3.20.3
psutil==5.9.8
pyaml==23.12.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pygame==2.1.0
PyOpenGL==3.1.7
pyparsing==3.1.2
pytest==8.1.1
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
pytype==2021.8.11
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rlax==0.1.6
rlds==0.1.8
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.2
scikit-optimize==0.10.1
scipy==1.12.0
six==1.16.0
SQLAlchemy==2.0.30
statsmodels==0.14.1
tabulate==0.9.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.15.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.4.0
tifffile==2024.2.12
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
trfl==1.2.0
typed-ast==1.5.5
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
